\name{eValuesCrpsAnalysis-package}
\alias{eValuesCrpsAnalysis-package}
\alias{eValuesCrpsAnalysis}
\docType{package}
\title{
\packageTitle{eValuesCrpsAnalysis}
}
\description{
\packageDescription{eValuesCrpsAnalysis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{eValuesCrpsAnalysis}
\packageIndices{eValuesCrpsAnalysis}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{eValuesCrpsAnalysis}

Maintainer: \packageMaintainer{eValuesCrpsAnalysis}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
